requirements.txt
weaviate-client==4.13.2
tiktoken==0.9.0
openai==1.75.0
langchain==0.3.24
langchain-weaviate==0.0.4
langchain-openai==0.3.14
langchain-community==0.3.22
pandas==2.2.3